refactor: 패키지 이동 및 랜덤 인터페이스 추가
