+++ backend/src/test/java/coffeeshout/coffeeshout/domain/service/MiniGamePlayServiceTest.java
@@ -7,7 +7,8 @@ import coffeeshout.coffeeshout.domain.JoinCode;
 import coffeeshout.coffeeshout.domain.Menu;
 import coffeeshout.coffeeshout.domain.MiniGame;
 import coffeeshout.coffeeshout.domain.Room;
-import coffeeshout.coffeeshout.domain.Roulette;
+import coffeeshout.coffeeshout.domain.roulette.JavaRandomGenerator;
+import coffeeshout.coffeeshout.domain.roulette.Roulette;
 import coffeeshout.coffeeshout.domain.player.Player;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
@@ -17,7 +18,7 @@ class MiniGamePlayServiceTest {
 
     private MiniGamePlayService service = new MiniGamePlayService();
     private List<Player> players = List.of(new Player(null, "꾹", null, null), new Player(null, "한스", null, null));
-    private Room room = new Room(new JoinCode("AAAAAA"), new Roulette(players, 1));
+    private Room room = new Room(new JoinCode("AAAAAA"), new Roulette(players, 1, new JavaRandomGenerator()));
     private Player host = new Player(1L, "host", new Menu(), room);
     private Player guest = new Player(2L, "guest", new Menu(), room);
